image collection: new commitments part 2
