+++ image_collection/data_extraction_tools.py
@@ -1,5 +1,7 @@
-## Webscraping Tool maangement 
+## Data Extraction Tools
 # 
+# purpose: extracting content from web resources 
+#
 # function 1: scraping images 
 # function 2: scraping labels 
 # function 3: scraping prices 
@@ -18,6 +20,17 @@ import pandas as pd
 import os 
 import requests
 import time
+
+# (WEBLINKS): implement function 0 --> extracting weblinks from resource files
+# extract weblinks from loaded resource file
+def extract_weblinks(filename:str) -> list: 
+  # define list --> store the weblink
+  weblink_list = []
+  # open the resource file
+  with open(filename, "r") as infile:
+    for link in infile: 
+      weblink_list.append(link)
+  return weblink_list
 
 # implement function for scraping images 
 # (IMAGEs): implement function 1 --> extracting images from url (CHECK) (Option 1)
@@ -41,7 +54,7 @@ def scrape_mobile_images(url: str) -> list:
     browser.close()
     return image_urls
 
-# extract images from web links (Option 2)
+# (IMAGES) function 1 --> extract images from web links (Option 2)
 def extract_mobile_images(weblink: str) -> list:
  # Set up Chrome options for headless mode and pop-up disabling
   chrome_options = Options()
@@ -120,93 +133,8 @@ def extract_mobile_price(url: str) -> list:
     return price_urls
     browser.close()
 
-# implement function 5: downloading images from image urls after extracting it from weblink (CHECK)
-def download_images(image_urls: list, 
-                    image_folder_dir: str): 
-  try: 
-    # create a directory to save images
-      os.makedirs(image_folder_dir, exist_ok=True)
-  except FileExistsError as file_exist_error:
-     print(f"Type of Error: {FileExistsError}")
-     print(f"Description: {file_exist_error}")
-     print(f"Filename: {file_exist_error.filename}")
 
-  # implement iteration: download each image  
-  for i, url in enumerate(image_urls):
-
-    if not url.startswith(("http://", "https://")):
-      print(f"Skipping invalid URL: {url}")
-      continue 
-    
-    try:
-        # Fetch the image
-        response = requests.get(url, stream=True)
-        response.raise_for_status()  # Check for HTTP errors
-
-        # Create a file name based on the image URL
-        file_extension = url.split('.')[-1]
-        filename = f"image_{i + 1}.{file_extension}"
-        filepath = os.path.join(image_folder_dir, filename)
-
-        # Save the image
-        with open(filepath, 'wb') as file:
-            for chunk in response.iter_content(1024):
-                file.write(chunk)
-        
-        print(f"Downloaded: {filename}")
-
-    except requests.exceptions.RequestException as e:
-        print(f"Failed to download {url}: {e}")
-        # image_urls.remove(url)
 
 ## Problem 
 # function 5: unknown url causes infinite looping --> remove unknown url# 
 
-# function  5: removing invalid urls
-def remove_invalid_urls(filename: str) -> list: 
-  # define 2 validated prefixes: http:// and https://
-  valid_url_prefix = ("http://","https://")
-
-  # open file: look for invalid urls for removal
-  with open(filename, "r") as infile: 
-    urls = [url.strip() for url in infile.readlines()]
-
-  # filter valid urls 
-  validate_urls = [url.strip() for url in urls if url.strip().startswith(valid_url_prefix)]
-  # overwrite current file by new list of urls
-  with open(filename, "w") as outfile:
-    for url in validate_urls:
-      outfile.write(url + "\n")
-
-  # compute number of valid and invalid urls
-  num_val_url = len(validate_urls)
-  num_inval_url = len(urls) - len(validate_urls)
-  print(f"Removed {len(urls) - len(validate_urls)} invalid urls")
-  return num_val_url, num_inval_url
-  return validate_urls
-
-
-# function 6: validate urls --> valid or invalid (CHECK)
-def validate_urls(filename:str) -> tuple:
-  # define list of urls: validated or invalidated
-  validate_urls = []
-  invalidated_urls = []
-  num_val_url, num_inval_url = (0,0)
-  # define prefix set for validated urls
-  valid_url_prefix = (("http://","https://"))
-
-  # open file --> check every url (valid + invalid)
-  with open(filename, "r") as infile:
-    for weburl in infile:
-      weburl.strip()  # removing whitespaces
-      if not weburl.startswith(valid_url_prefix): 
-        invalidated_urls.append(weburl)
-        num_inval_url += 1
-      else:
-        validate_urls.append(weburl)
-        num_val_url += 1
-    
-    return num_val_url, num_inval_url
-          
-
-
